# docs/swagger.yaml
basePath: /v1
definitions:
  models.Expense:
    properties:
      amount:
        type: number
      buyer:
        type: string
      date:
        type: string
      description:
        type: string
      expenseListId:
        type: string
      id:
        type: string
      participants:
        items:
          type: string
        type: array
    type: object
  models.ExpenseList:
    properties:
      color:
        type: string
      creatorId:
        type: string
      currency:
        type: string
      emoji:
        type: string
      expenses:
        items:
          $ref: '#/definitions/models.Expense'
        type: array
      id:
        type: string
      inviteCode:
        type: string
      participants:
        items:
          $ref: '#/definitions/models.Participant'
        type: array
      title:
        type: string
      totalCost:
        type: number
    type: object
  models.ExpenseListCompensation:
    properties:
      amount:
        type: number
      from:
        type: string
      to:
        type: string
    type: object
  models.ExpenseListShare:
    properties:
      difference:
        type: number
      expenseAmount:
        type: number
      id:
        type: string
      numberOfExpenses:
        type: integer
      share:
        type: number
    type: object
  models.ExpenseListWrapper:
    properties:
      compensations:
        items:
          $ref: '#/definitions/models.ExpenseListCompensation'
        type: array
      expenseList:
        $ref: '#/definitions/models.ExpenseList'
      shares:
        items:
          $ref: '#/definitions/models.ExpenseListShare'
        type: array
    type: object
  models.Participant:
    properties:
      avatarUrl:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: This is the backend for the equaly cost management app
  title: Equaly Backend
  version: "1.0"
paths:
  /expense:
    post:
      description: Creates a new Expense for an Expense List
      parameters:
      - description: Expense Data
        in: body
        name: expense
        required: true
        schema:
          $ref: '#/definitions/models.Expense'
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Updated Expense List with new Expense, Compensation and Shares
          schema:
            $ref: '#/definitions/models.ExpenseListWrapper'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create Expense
      tags:
      - Expenses
  /expense-list:
    post:
      description: Creates a new Expense list with given data for a specified user
      parameters:
      - description: Expense List Data
        in: body
        name: expenseList
        required: true
        schema:
          $ref: '#/definitions/models.ExpenseList'
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Created Expense List
          schema:
            $ref: '#/definitions/models.ExpenseListWrapper'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create Expense List
      tags:
      - Expenses
  /expense-list/join:
    post:
      description: Join ExpenseList with valid inviteCode
      parameters:
      - description: Expense List Id
        in: query
        name: expenseListId
        required: true
        type: string
      - description: Invite Code
        in: query
        name: inviteCode
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Join Expense List
      tags:
      - Expenses
  /expense-lists:
    get:
      description: Retrieves the list of expenses for a specified user.
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: List of user expense lists
          schema:
            items:
              $ref: '#/definitions/models.ExpenseListWrapper'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get Expense Lists
      tags:
      - Expenses
  /static/profile/{expenseListId}/{participantId}:
    get:
      description: Retrieves presigned url of profile picture
      parameters:
      - description: Expense List Id
        in: path
        name: expenseListId
        required: true
        type: string
      - description: Participant Id
        in: path
        name: participantId
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: "Success\"\tstring\t\"Profile Picture presigned URL"
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Get Profile Picture presigned Url
      tags:
      - Expenses
    post:
      description: Upload Profile Picture of Participant
      parameters:
      - description: Expense List Id
        in: path
        name: expenseListId
        required: true
        type: string
      - description: Participant Id
        in: path
        name: participantId
        required: true
        type: string
      - description: "Profile\tPicture"
        in: formData
        name: image
        required: true
        type: file
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: Success
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Upload Profile Picture of Participant
      tags:
      - Expenses
swagger: "2.0"